fix(parsing): fixed bug with exp/fame numbers being bigger than int max val

+++ RealmAspNet/RealmEye/Definitions/Player/PlayerData.cs
@@ -6,8 +6,8 @@ namespace RealmAspNet.RealmEye.Definitions.Player
 	{
 		public int CharacterCount { get; set; }
 		public int Skins { get; set; }
-		public int Fame { get; set; }
-		public int Exp { get; set; }
+		public long Fame { get; set; }
+		public long Exp { get; set; }
 		public int Rank { get; set; }
 		public int AccountFame { get; set; }
 		public string Guild { get; set; }
@@ -26,7 +26,7 @@ namespace RealmAspNet.RealmEye.Definitions.Player
 		public string CharacterType { get; init; }
 		public int Level { get; init; }
 		public int ClassQuestsCompleted { get; init; }
-		public int Fame { get; init; }
+		public long Fame { get; init; }
 		public long Experience { get; init; }
 		public int Place { get; init; }
 		public GearInfo[] EquipmentData { get; init; }
